proper kill

+++ stop.sh
@@ -11,7 +11,9 @@ pushd "$HOME/lnxdiagd"
     # command the daemon to stop regardless if it is running or not.
     eval "./lnxdiag$daemon"d.py stop
     # kill off any rogue daemons by the same name (it happens sometimes)
-    kill $(pgrep -f "lnxdiag$daemon"d.py)
+    if [   $(pgrep -fc "lnxdiag$daemon"d.py) -ne 0 ]; then
+      kill $(pgrep -f  "lnxdiag$daemon"d.py)
+    fi
     # log the activity
     logger -p user.err -t lnxdiagd "  * Daemon $daemon Stopped."
     # force rm the .pid file
@@ -23,7 +25,9 @@ pushd "$HOME/lnxdiagd"
     # command the daemon to stop regardless if it is running or not.
     eval "./lnxsvc$daemon"d.py stop
     # kill off any rogue daemons by the same name (it happens sometimes)
-    kill $(pgrep -f "lnxdiag$daemon"d.py)
+    if [   $(pgrep -fc "lnxsvc$daemon"d.py) -ne 0 ]; then
+      kill $(pgrep -f  "lnxsvc$daemon"d.py)
+    fi
     # log the activity
     logger -p user.err -t lnxdiagd "  * Daemon $daemon Stopped."
     # force rm the .pid file
